move click handlers to utils
