+++ src/lib.rs
@@ -368,8 +368,6 @@ impl Plugin for CosmicEditPlugin {
             .add_systems(
                 Update,
                 (
-                    change_active_editor_ui,
-                    change_active_editor_sprite,
                     cosmic_edit_bevy_events,
                     cosmic_edit_set_redraw,
                     on_scale_factor_change,
@@ -392,57 +390,6 @@ impl Plugin for CosmicEditPlugin {
 #[derive(Resource, Default)]
 pub struct ActiveEditor {
     pub entity: Option<Entity>,
-}
-
-fn change_active_editor_ui(
-    mut commands: Commands,
-    mut interaction_query: Query<
-        (&Interaction, Entity),
-        (
-            Changed<Interaction>,
-            (With<CosmicEditor>, Without<ReadOnly>),
-        ),
-    >,
-) {
-    for (interaction, entity) in interaction_query.iter_mut() {
-        if let Interaction::Pressed = interaction {
-            commands.insert_resource(ActiveEditor {
-                entity: Some(entity),
-            });
-        }
-    }
-}
-
-fn change_active_editor_sprite(
-    mut commands: Commands,
-    windows: Query<&Window, With<PrimaryWindow>>,
-    buttons: Res<Input<MouseButton>>,
-    mut cosmic_edit_query: Query<
-        (&mut Sprite, &GlobalTransform, Entity),
-        (With<CosmicEditor>, Without<ReadOnly>),
-    >,
-    camera_q: Query<(&Camera, &GlobalTransform)>,
-) {
-    let window = windows.single();
-    let (camera, camera_transform) = camera_q.single();
-    if buttons.just_pressed(MouseButton::Left) {
-        for (sprite, node_transform, entity) in &mut cosmic_edit_query.iter_mut() {
-            let size = sprite.custom_size.unwrap_or(Vec2::new(1., 1.));
-            let x_min = node_transform.affine().translation.x - size.x / 2.;
-            let y_min = node_transform.affine().translation.y - size.y / 2.;
-            let x_max = node_transform.affine().translation.x + size.x / 2.;
-            let y_max = node_transform.affine().translation.y + size.y / 2.;
-            if let Some(pos) = window.cursor_position() {
-                if let Some(pos) = camera.viewport_to_world_2d(camera_transform, pos) {
-                    if x_min < pos.x && pos.x < x_max && y_min < pos.y && pos.y < y_max {
-                        commands.insert_resource(ActiveEditor {
-                            entity: Some(entity),
-                        });
-                    };
-                }
-            };
-        }
-    }
 }
 
 /// Resource struct that holds configuration options for cosmic fonts.
@@ -965,11 +912,7 @@ pub fn cosmic_edit_bevy_events(
                 if let Some(node_cursor_pos) = get_node_cursor_pos(
                     primary_window,
                     node_transform,
-                    (
-                        // TODO this does not work with percent sizes
-                        // well like I haven't tested it so it might
-                        width, height,
-                    ),
+                    (width, height),
                     is_ui_node,
                     camera,
                     camera_transform,
